pyproject.toml
[tool.poetry]
name = "andelaskillsseries-ai-ml"
version = "0.1.0"
description = ""
authors = ["Bertan ULUSOY <bertan_ulusoy@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.13.3"
langchain = "^0.1.11"
pinecone-client = "^3.1.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.6.0"
docx2txt = "^0.8"
pypdf = "^4.1.0"
requests = "^2.31.0"
pandas = "^2.2.1"
aiohttp = "^3.9.3"
pyyaml = "^6.0.1"
numpy = "^1.26.4"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
